tech(env) - docker for php application
