+++ composer.json
@@ -13,7 +13,7 @@
     },
     "autoload": {
         "psr-4": {
-            "jdRoll\\": "src/"
+            "jdRoll\\": "src/jdRoll/"
         }
     }
 }
