+++ docker-compose.yml
@@ -1,15 +1,5 @@
 version: '3.3'
 services:
-  portainer:
-    image: portainer/portainer
-    command: -H unix:///var/run/docker.sock
-    restart: always
-    ports:
-      - 9000:9000
-      - 8000:8000
-    volumes:
-      - /var/run/docker.sock:/var/run/docker.sock
-      - portainer_data:/data
   db:
     image: mysql:5.7
     restart: always
@@ -18,6 +8,7 @@ services:
       MYSQL_USER: 'jdroll'
       MYSQL_PASSWORD: 'devdev'
       MYSQL_ROOT_PASSWORD: 'devdev'
+    command: --sql_mode=""
     ports:
       - '3306:3306'
     expose:
@@ -25,6 +16,8 @@ services:
     volumes:
       - jdroll-db:/var/lib/mysql
       - ./docker/init.sql:/docker-entrypoint-initdb.d/init.sql:ro
+      - ./docker/data.sql:/sql/data.sql:ro
+      - ./docker/load-data.sh:/sql/load-data.sh:ro
   phpmyadmin:
     image: phpmyadmin/phpmyadmin
     container_name: pma
@@ -37,6 +30,21 @@ services:
     restart: always
     ports:
       - 8081:80
+  php:
+    build: ./
+    links:
+      - db
+    volumes:
+      - ./php.ini:/usr/local/etc/php/php.ini
+      - ./:/code
+  web:
+    image: nginx:1.19.8-alpine
+    links:
+      - php
+    volumes:
+      - ./web.conf:/etc/nginx/conf.d/default.conf
+      - ./:/code
+    ports:
+      - "8080:80"
 volumes:
-  portainer_data:
   jdroll-db:
