+++ docker/load-data.sh
@@ -1,0 +1,1 @@
+cat /sql/data.sql | mysql -h localhost -P 3306 -u jdroll -pdevdev -D jdroll --protocol=tcp
